move utils for its own file
In the server was getting an import error.
